components/contact/form/fields/message-field.tsx
"use client"

import { FormField } from "@/components/ui/form"
import { Textarea } from "@/components/ui/textarea"
import { UseFormReturn } from "react-hook-form"
import { ContactFormData } from "../contact-form-schema"

interface MessageFieldProps {
  form: UseFormReturn<ContactFormData>
}

export function MessageField({ form }: MessageFieldProps) {
  return (
    <FormField
      control={form.control}
      name="message"
      render={({ field }) => (
        <FormItem>
          <FormLabel>Message</FormLabel>
          <FormControl>
            <Textarea 
              placeholder="Tell us more about your inquiry..."
              className="min-h-[150px]"
              {...field}
            />
          </FormControl>
          <FormMessage />
        </FormItem>
      )}
    />
  )
}